requirements.txt
streamlit>=1.28.0
python-telegram-bot>=20.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
redis>=5.0.0
